Add glibc_malloc_max_trim_threshold GUC.
This GUC allows the user to dynamically change the max trim threshold
used in calls to mallopt.

When set to the default -1 value, we cap it to 16MB since it is two
times the most common max block size, or to work_mem / 2 since for
values less than 16MB we won't allocate blocks bigger than 8MB.

When set to another value, we cap it to work_mem.