Use a non-blocking call for archive_command
Instead of using system directly, use popen (through the OpenPipeStream
abstraction) and process interrupts while the command is running.

The fact that the archiver process did not process interrupts in the
past was not a big problem, but since the archiver process now alsot
must acknowledge signals for ProcSignalBarriers to be lifted, a slow
archive_command can unnecessarily block other processes.